Move digital to own component
