+++ src/panels/lovelace/cards/hui-clock-card.ts
@@ -1,4 +1,3 @@
-import type { PropertyValues } from "lit";
 import { css, html, LitElement, nothing } from "lit";
 import { customElement, property, state } from "lit/decorators";
 import { classMap } from "lit/directives/class-map";
@@ -11,10 +10,9 @@ import type {
   LovelaceGridOptions,
 } from "../types";
 import type { ClockCardConfig } from "./types";
-import { useAmPm } from "../../../common/datetime/use_am_pm";
-import { resolveTimeZone } from "../../../common/datetime/resolve-time-zone";
+import "./clock/hui-clock-card-digital";
 
-const INTERVAL = 1000;
+export const INTERVAL = 1000;
 
 @customElement("hui-clock-card")
 export class HuiClockCard extends LitElement implements LovelaceCard {
@@ -33,45 +31,8 @@ export class HuiClockCard extends LitElement implements LovelaceCard {
 
   @state() private _config?: ClockCardConfig;
 
-  @state() private _dateTimeFormat?: Intl.DateTimeFormat;
-
-  @state() private _timeHour?: string;
-
-  @state() private _timeMinute?: string;
-
-  @state() private _timeSecond?: string;
-
-  @state() private _timeAmPm?: string;
-
-  private _tickInterval?: undefined | number;
-
   public setConfig(config: ClockCardConfig): void {
     this._config = config;
-    this._initDate();
-  }
-
-  private _initDate() {
-    if (!this._config || !this.hass) {
-      return;
-    }
-
-    let locale = this.hass?.locale;
-
-    if (this._config?.time_format) {
-      locale = { ...locale, time_format: this._config.time_format };
-    }
-
-    this._dateTimeFormat = new Intl.DateTimeFormat(this.hass.locale.language, {
-      hour: "2-digit",
-      minute: "2-digit",
-      second: "2-digit",
-      hourCycle: useAmPm(locale) ? "h12" : "h23",
-      timeZone:
-        this._config?.time_zone ||
-        resolveTimeZone(locale.time_zone, this.hass.config?.time_zone),
-    });
-
-    this._tick();
   }
 
   public getCardSize(): number {
@@ -109,50 +70,6 @@ export class HuiClockCard extends LitElement implements LovelaceCard {
     };
   }
 
-  protected updated(changedProps: PropertyValues) {
-    if (changedProps.has("hass")) {
-      const oldHass = changedProps.get("hass");
-      if (!oldHass || oldHass.locale !== this.hass?.locale) {
-        this._initDate();
-      }
-    }
-  }
-
-  public connectedCallback() {
-    super.connectedCallback();
-    this._startTick();
-  }
-
-  public disconnectedCallback() {
-    super.disconnectedCallback();
-    this._stopTick();
-  }
-
-  private _startTick() {
-    this._tickInterval = window.setInterval(() => this._tick(), INTERVAL);
-    this._tick();
-  }
-
-  private _stopTick() {
-    if (this._tickInterval) {
-      clearInterval(this._tickInterval);
-      this._tickInterval = undefined;
-    }
-  }
-
-  private _tick() {
-    if (!this._dateTimeFormat) return;
-
-    const parts = this._dateTimeFormat.formatToParts();
-
-    this._timeHour = parts.find((part) => part.type === "hour")?.value;
-    this._timeMinute = parts.find((part) => part.type === "minute")?.value;
-    this._timeSecond = this._config?.show_seconds
-      ? parts.find((part) => part.type === "second")?.value
-      : undefined;
-    this._timeAmPm = parts.find((part) => part.type === "dayPeriod")?.value;
-  }
-
   protected render() {
     if (!this._config) return nothing;
 
@@ -170,16 +87,10 @@ export class HuiClockCard extends LitElement implements LovelaceCard {
           ${this._config.title !== undefined
             ? html`<div class="time-title">${this._config.title}</div>`
             : nothing}
-          <div class="time-parts">
-            <div class="time-part hour">${this._timeHour}</div>
-            <div class="time-part minute">${this._timeMinute}</div>
-            ${this._timeSecond !== undefined
-              ? html`<div class="time-part second">${this._timeSecond}</div>`
-              : nothing}
-            ${this._timeAmPm !== undefined
-              ? html`<div class="time-part am-pm">${this._timeAmPm}</div>`
-              : nothing}
-          </div>
+          <hui-clock-card-digital
+            .hass=${this.hass}
+            .config=${this._config}
+          ></hui-clock-card-digital>
         </div>
       </ha-card>
     `;
@@ -233,74 +144,6 @@ export class HuiClockCard extends LitElement implements LovelaceCard {
     .time-wrapper.size-large .time-title {
       font-size: var(--ha-font-size-2xl);
       line-height: var(--ha-line-height-condensed);
-    }
-
-    .time-parts {
-      align-items: center;
-      display: grid;
-      grid-template-areas:
-        "hour minute second"
-        "hour minute am-pm";
-
-      font-size: 2rem;
-      font-weight: var(--ha-font-weight-medium);
-      line-height: 0.8;
-      direction: ltr;
-    }
-
-    .time-title + .time-parts {
-      font-size: 1.5rem;
-    }
-
-    .time-wrapper.size-medium .time-parts {
-      font-size: 3rem;
-    }
-
-    .time-wrapper.size-large .time-parts {
-      font-size: 4rem;
-    }
-
-    .time-wrapper.size-medium .time-parts .time-part.second,
-    .time-wrapper.size-medium .time-parts .time-part.am-pm {
-      font-size: var(--ha-font-size-l);
-      margin-left: 6px;
-    }
-
-    .time-wrapper.size-large .time-parts .time-part.second,
-    .time-wrapper.size-large .time-parts .time-part.am-pm {
-      font-size: var(--ha-font-size-2xl);
-      margin-left: 8px;
-    }
-
-    .time-parts .time-part.hour {
-      grid-area: hour;
-    }
-
-    .time-parts .time-part.minute {
-      grid-area: minute;
-    }
-
-    .time-parts .time-part.second {
-      grid-area: second;
-      line-height: 0.9;
-      opacity: 0.4;
-    }
-
-    .time-parts .time-part.am-pm {
-      grid-area: am-pm;
-      line-height: 0.9;
-      opacity: 0.6;
-    }
-
-    .time-parts .time-part.second,
-    .time-parts .time-part.am-pm {
-      font-size: var(--ha-font-size-xs);
-      margin-left: 4px;
-    }
-
-    .time-parts .time-part.hour:after {
-      content: ":";
-      margin: 0 2px;
     }
   `;
 }
